Add Stars to Rating
Create a folder of style inside src to contain all sass files. Also,
move App.css into src/styles because this will be where we compile
all of our sass into css.

Add all sass files for fontawesome into styles and the fonts into
public/fonts. We will need this to render stars for ratings.

Update path to font files for font-awesome to load all fonts in it's
style sheets. Run sass watch to watch for changes and compile it all
too App.css, this will generate a map as well.

Update beauty place component with a beauty place header to wrap the
name and rating.

Add star components and logic to render full, half and empty stars
depending on the rating.

Create components for empty star, full and half star.

Add style to make the beauty place name an inline element so the
star rating can be inline with the name.