Refactor Stateful Components To Be Stateless
Delete a component that's not being used BeautyPlaces
becaause we will use BeautyPlaceList instead.

Create a folder containers with a file BeautyPlaceListContainer.
This will be responsible for defining any kind of behavior we will use
and allow us to make the BeautyPlacesList and all children stateless.

Remove unused components and refactor as many stateful
components into stateless components in order to simplify our code.