Update Update Location Form
remove Searchable from SearchableBeautyPlacesContainer to simplify name
since we only really need to know this a container for beauty_places
and therefore it's only responsibility is maintaining state.

Change loc to locationSuggestion in searchform component because
we needed it to be more descriptive. Add currentLocation to state
from the props attribute. Add a set state call for current location
inside of the component did mount hook.

Add a function definition for setCurrentLocation in the beauty place
container and set it's currentlocation in order to cache location.

Add fetch current location service call to beauty place service

Update sass for the search form to look good.