Add Redux
Abstract all state and behavior into actions from the
beauty places container in order to enforce
separation of concerns.

Add the provider and store in order to maintain state
throughout a container and it's children components.

Update the url for fetchByCityState service since we
want to nest the city state values in a current_location
hash for passing it to the api.

Create actions reducers and store in order to enforce
seperation of concerns and mutate state.