Update Beauty Places Search Form
Add an action to support when a user changed a location
to and add a handler for that to pass a message to the
corresponding reducer in order to update currentLocation
in our redux store..

Update form to succesfully submit a users input to change
the current location.

Add logic to container that will check if there is userInput
for the currentLocation and if there is it will use skip
fetching current location based off of coord position.

Add user input key value to the default state in the
CurrentLocation reducer.

Add service that fetches beauty places based off of
the user input.

Update some styles to make search form pretty. :)

Change this.state.placeId to this.props.placeId since state
is asynchronous and it would never set in time for render. It's
better to just use props instead of lifecycle hooks.

Update some styles